+++ Gemfile
@@ -44,6 +44,9 @@ gem 'vault'
 gem 'rollbar'
 gem 'webmock', '>= 3.12.1'
 
+# aws ssm
+gem 'aws-sdk-ssm'
+
 group :development, :test do
   gem 'listen', '~> 3.2'
   # Spring speeds up development by keeping your application running in the background. Read more: https://github.com/rails/spring
