+++ config/initializers/cf_vault.rb
@@ -1,5 +1,3 @@
-require 'rollbar'
-
 def config_vault
   vcap_services = JSON.parse(ENV['VCAP_SERVICES'])
 
@@ -23,18 +21,6 @@ def set_env(storage_path)
   env_vars.data.each do |env_key, env_value|
     ENV[env_key.to_s] = env_value.to_s
   end
-  config_rollbar
 end
 
-def config_rollbar
-  Rollbar.configure do |config|
-    ['sandbox'].each do |env|
-      config.access_token = ENV['ROLLBAR_ACCESS_TOKEN']
-      config.environment = env
-    end
-  end
-  Rails.logger.info('App Deployed & Rollbar Successfully Configured')
-  Rollbar.info('App Deployed & Rollbar Successfully Configured')
-end
-
-config_vault if ENV['SERVER_ENV_NAME'].present?
+# config_vault if ENV['SERVER_ENV_NAME'].present?
